requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt setup.py
#
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
decorator==4.4.2
    # via moviepy
filelock==3.16.1
    # via torch
fsspec==2024.10.0
    # via torch
idna==3.10
    # via requests
imageio==2.36.0
    # via moviepy
imageio-ffmpeg==0.5.1
    # via moviepy
jinja2==3.1.4
    # via torch
llvmlite==0.43.0
    # via numba
markupsafe==3.0.2
    # via jinja2
more-itertools==10.5.0
    # via openai-whisper
moviepy==1.0.3
    # via open_video_transcriber (setup.py)
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==2.0.2
    # via
    #   imageio
    #   moviepy
    #   numba
    #   open_video_transcriber (setup.py)
    #   openai-whisper
openai-whisper==20240930
    # via open_video_transcriber (setup.py)
pillow==11.0.0
    # via imageio
proglog==0.1.10
    # via moviepy
pyqt5==5.15.11
    # via open_video_transcriber (setup.py)
pyqt5-qt5==5.15.15
    # via pyqt5
pyqt5-sip==12.15.0
    # via pyqt5
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   moviepy
    #   tiktoken
sympy==1.13.3
    # via torch
tiktoken==0.8.0
    # via openai-whisper
torch==2.2.2
    # via
    #   open_video_transcriber (setup.py)
    #   openai-whisper
tqdm==4.66.5
    # via
    #   moviepy
    #   openai-whisper
    #   proglog
typing-extensions==4.12.2
    # via torch
urllib3==2.2.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
